backend/pyproject.toml
[tool.poetry]
name = "news-mlops-free-backend"
version = "0.1.0"
description = "Backend for the News MLOps Free project."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pymongo = "^4.6.0"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
newspaper3k = "^0.2.8"
transformers = "^4.35.2"
torch = "^2.1.0" # Use CPU version for smaller size, adjust if GPU is needed
scikit-learn = "^1.3.2"
requests = "^2.31.0"
apscheduler = "^3.10.4"
mlflow = "^2.8.1"
gunicorn = "^21.2.0"
dvc = {extras = ["s3"], version = "^3.33.0"} # Example with S3 remote

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
